Add a way to generate demo page (#17)
Add a `build-demo` script which will build the demo page to the
`demo-build/` directory. This directory can then be copied to a
`gh-pages` branch so that we can publish the demo page to GitHub Pages.

Also, considering that we also need an HTML file to build the demo page,
move that along with the JavaScript file to the `demo/` directory.